+++ ChangeLog.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add support for contract schemas containing the version.
+- Add support for the contract schema types, ULeb128, ILeb128, ByteList and ByteArray.
+
 ## 4.0.3
 
 - Fix calculation of payload size of a configure baker transaction in
@@ -9,8 +12,6 @@
   the wallet-proxy.
 - Fix help text of `baker add` and `baker configure` commands that referred to
   `raw GetChainParameters`.
-- Add support for contract schemas containing version.
-- Add support for the contract schema types, ULeb128, ILeb128, ByteList and ByteArray.
 
 ## 4.0.2